fix the warnings

+++ lib/encode_impl.cc
@@ -573,7 +573,7 @@ namespace gr {
             }
           }
           pmt::pmt_t pairs = pmt::dict_items(dict);
-          for (int i = 0; i < pmt::length(pairs); i++) {
+          for (size_t i = 0; i < pmt::length(pairs); i++) {
               pmt::pmt_t pair = pmt::nth(i, pairs);
               add_item_tag(0,                   // output port index
                             nitems_written(0),  // output sample index
